fix issue when packets are sent/received by Sancus at the same time
This happens when:
- sancus is sending a RemoteOutput packet
- at the same time, a command is sent to sancus

when the command is sent to sancus, the handshake fails (because what we 
read from the UART is not the handshake response but the RemoteOutput 
packet). 

To solve this, we need to "pause" the handshake and handle the incoming 
packet instead. After that, we try again to read the handshake response 
and, if there are no more incoming packets, we continue with the 
transmission to Sancus